sentineliq/api/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
imbalanced-learn
sentence-transformers
faiss-cpu
anthropic